fix: utils
